requirements.txt
absl-py==0.9.0
amqp==2.5.2
astor==0.8.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels_redis
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
djangorestframework==3.11.0
ffmpeg-python==0.2.0
future==0.18.2
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
incremental==17.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==4.6.8
Markdown==3.1.1
msgpack==0.6.0
mysqlclient==1.4.6
norbert==0.2.1
numpy==1.18.1
pandas==0.25.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.2.0
requests==2.22.0
scipy==1.4.1
service-identity==18.1.0
six==1.14.0
spleeter==1.4.9
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.15.2
tensorflow-estimator==1.14.0
termcolor==1.1.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.0.0
zope.interface==4.7.1
